requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
cx-Oracle==8.3.0
fastapi==0.88.0
h11==0.14.0
httptools==0.5.0
idna==3.4
Pillow==9.3.0
pydantic==1.10.2
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
